move cart to its own subroute, clean up api route
